% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloomberg-data-point.R
\name{bloomberg_data_point}
\alias{bloomberg_data_point}
\title{Function getting a single data point}
\usage{
bloomberg_data_point(tickers = "ADS GY", type = "Equity",
  fields = "PX_LAST", option.fields = NULL, option.values = NULL,
  override.fields = NULL, override.values = NULL)
}
\arguments{
\item{tickers}{A character vector specifying the securities}

\item{type}{A character string specifying instrument type (optional)}

\item{fields}{A character vector specifying data fields}

\item{override.values}{character vector specifying override values}

\item{override.names}{character vector specifying override fields}
}
\description{
Function getting a single data point
}
\author{
Peter Garnry
}

