% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloomberg-earnings-estimates.R
\name{bloomberg_earnings_estimates}
\alias{bloomberg_earnings_estimates}
\title{Function getting current Bloomberg estimates}
\usage{
bloomberg_earnings_estimates(tickers = "GS US", type = "Equity",
  fields = "BEST_EPS", currency = NULL)
}
\arguments{
\item{tickers}{character vector specifying the securities}

\item{type}{a character string specifying instrument type
price point extraction}

\item{fields}{character vector specifying data fields
price point extraction (see details)}

\item{currency}{a character string specifying the currency}
}
\description{
Function getting current Bloomberg estimates
}
\details{
The function finds and implement the company's
primary periodicity so the company's estimates are either
quarterly, semi-annually or annually depending on the
available period.
}

