% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloomberg-historical-data.R
\name{bloomberg_historical_data}
\alias{bloomberg_historical_data}
\title{Function getting historical Bloomberg data}
\usage{
bloomberg_historical_data(tickers = "GS US", type = "Equity",
  fields = "PX_LAST", freq = "DAILY", start.date = "2014-01-01",
  end.date = "", currency = NULL,
  non.trading.days = "NON_TRADING_WEEKDAYS",
  non.trading.days.fill = "NIL_VALUE", calendar.type = "CALENDAR",
  pricing = "PRICING_OPTION_PRICE", follow.dpdf = TRUE,
  override.fields = NULL, override.values = NULL)
}
\arguments{
\item{tickers}{character vector specifying the securities}

\item{type}{a character string specifying instrument type
price point extraction}

\item{fields}{character vector specifying data fields
price point extraction (see details)}

\item{freq}{a character string specifying time frequency}

\item{currency}{a character string specifying the currency}
}
\description{
Function getting historical Bloomberg data
}
\details{
Historical data on monthly and quarterly frequency
are exported with Index Class 'yearmon' and 'yearqtr'
respectively as tickers across various markets do have the
same month end dates. Using this methodology ensure that
data is aligned properly.
}

