% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloomberg-historical-index-members.R
\name{bloomberg_historical_index_members}
\alias{bloomberg_historical_index_members}
\title{Get historical index tickers to avoid surviourship bias
The function works for one equity index at a time}
\usage{
bloomberg_historical_index_members(index = "SPX", id = "ticker",
  start.date = "2014-01-01", end.date = "", freq = "MONTHLY")
}
\arguments{
\item{index}{a character string specifying the index ticker}

\item{freq}{a character string specifying time frequency}

\item{field}{specifying the output instrument identification}

\item{startdate}{a character string specifying the start date}

\item{enddate}{a character string specifying the end date}
}
\description{
Get historical index tickers to avoid surviourship bias
The function works for one equity index at a time
}
\details{
The field variable has two options: "ticker" or "isin" 
enabling the function to return the desired instrument identification.

The frequency variable is set using the standard Bloomberg inputs such 
as "MONTHLY", "DAILY" etc.
}

