% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloomberg-index-members.R
\name{bloomberg_index_members}
\alias{bloomberg_index_members}
\title{Function loading all tickers of a specified index}
\usage{
bloomberg_index_members(index = "SPX")
}
\arguments{
\item{index}{Bloomberg index ticker code}
}
\description{
Function loading all tickers of a specified index
}
\examples{
IndexMembers(index = "SPX")
}
\author{
Peter Garnry
}

