% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_point.R
\name{DataPoint}
\alias{DataPoint}
\title{Function getting a single data point}
\usage{
DataPoint(tickers = "ADS GY", type = "Equity", fields = "PX_LAST",
  option.names = NULL, option.values = NULL, override.names = NULL,
  override.values = NULL)
}
\arguments{
\item{tickers}{A character vector specifying the securities}

\item{type}{A character string specifying instrument type (optional)}

\item{fields}{A character vector specifying data fields}

\item{override.names}{character vector specifying override fields}

\item{override.values}{character vector specifying override values}
}
\description{
Function getting a single data point
}
\author{
Peter Garnry
}

